requirement.txt
requests
python-telegram-bot
wikipediaapi
